Fix variable undefined bug
Pull request #04 introduced a bug, where the variable B_done was renamed to Btn but not in all places. A simple search and replace would have prevented this.

This fixes that by changing the name back to B_done which was more consistent with the rest of the code and at the same time fixes the aforementioned bug.