Use appropriate xdg user directory
Files that are downloaded should not simply be dumped in the user home
directory. The appropriate location is the xdg user download directory.

See: https://www.freedesktop.org/wiki/Software/xdg-user-dirs/

Some Pros of new behavior:
- Many users collect a lot of files and much later clean it all at
  once. This is more manageable in a dedicated directory.
- Some applications, like dedicated file cleaning utilities expect files
  to be stored in certain locations.
- Further, automatic backups programs that exclude downloads are less
  likely to waste space on temporary media files.

This also fits the Flatpak sand-boxing model. With the old behavior
a user would find the files under `~.var/app/com.github.JannikHv.Gydl`,
which would be less user friendly. However Flatpak allows granting
access to specific xdg user directories, which would work well with this
change.

In the end a lot of people would probably prefer a user specified
location. However because of the planned rewrite, a smaller fix like this
that will most likely cover the majority of users is a good compromise.