Improve appdata & screenshot
Further improvments on the appdata that better comply with the strict
validation tests. With this change the only missing or incomplete tag
is the <translation> tag. But this cannot be added since the app does
not support translations.

- Add swedish translation of the metadata (provides reference)
- Add update contact (Flatpak maintainer for now)
- Add high resolution screenshot
- Add developer name

Improvments on the screenshot is mostly in terms of appdata
specification compliance. But it is now also guaranteed to always be
available by being served from the git repo.

Appdata compliance:
- Has an aspect ratio of 16:9 (close enough for the validator)
- Has a width well above 620px (taken on 4K display)
- Uses default themes
- Only covers the window (no background)

https://www.freedesktop.org/software/appstream/docs/chap-Quickstart.html#qsr-app-screenshots-info