Cleanup installation instructions
Don't recommend yaourt when there are several better alternatives. In
any case, no AUR helper is supported by the Arch Linux team and if one
is installed the user should already know how to use it, i.e. not the
responsibility of gydl docs.

An alternative to this change, would be mentioning that gydl can be
installed, using an AUR helper and linking to this wiki page:

- https://wiki.archlinux.org/index.php/AUR_helpers

Also ran a spell checker and made some grammar improvments, as well as
ensuring there was a consistent line break at 80 characters.