Add initial Meson support
Building and packaging Gydl is greatly simplified by using Meson.

Additionally makes it so much simpler to handle translations if those
were to be added.

The py extension was removed from the gydl script because I could not
find an easy way to install it with a different name. Installing it with
the extension would not be proper.